==> PART 1-Movies-IMDB.ipynb <==
import os
os.makedirs('Data/',exist_ok=True) # Confirm folder created
os.listdir("Data/")

import json
with open('/Users/patelmedzy/.secret/mysql.json') as f:
    login = json.load(f)
login.keys()


#IMPORTS
import pandas as pd
import numpy as np
from sqlalchemy import create_engine
from sqlalchemy_utils import database_exists
import pymysql
from urllib.parse import quote_plus as urlquote
import pymysql
pymysql.install_as_MySQLdb()
from sqlalchemy import create_engine
from sqlalchemy_utils import create_database, database_exists
# Create connection string using credentials following this format
#connection = "dialect+driver://username:password@host:port/database"
#connection_str = "mysql+pymysql://root:root@localhost/books"
#connection = f"mysql+pymysql://{login['username']}:{urlquote(login['password'])}@localhost/books"
#engine = create_engine(connection)

# Download several files from IMDB’s movie data set and filter out the subset of moves requested by the stakeholder.

!('https://www.themoviedb.org/assets/2/v4/logos/v2/blue_long_1-8ba2ac31f354005783fab473602c34c3f4fd207150182061e425d366e4f34596.svg')

## Dataset Info

### Basics

#IMPORT Title basics
basics_url="https://datasets.imdbws.com/title.basics.tsv.gz"
basics = pd.read_csv(basics_url, sep='\t', low_memory=False)


basics.head()

basics.info()

### Title Ratings

#Import Title ratings
ratings_url="https://datasets.imdbws.com/title.ratings.tsv.gz"
ratings = pd.read_csv(ratings_url, sep='\t', low_memory=False)
ratings.head()

ratings.info()

### AKAS

#Import Title akas
akas = pd.read_csv("Data/title-akas-us-only.csv", low_memory=False)
akas.head()

akas.info()

## Data Cleaning

### AKAS

akas.replace({'\\N': np.nan}, inplace=True)

akas.head()

akas['region'].value_counts()

akas.info()

### Basics

# Filter the basics table down to only include the US by using the filter akas dataframe
keepers = basics["tconst"].isin(akas["titleId"])
basics = basics[keepers]
basics

basics.replace({'\\N': np.nan}, inplace=True)

basics.head(2)

# Filter movies released between 2000 and 2021 (inclusive)
basics = basics[(basics['startYear'].astype(float) >= 2000) & (basics['startYear'].astype(float) <= 2021)]



basics['startYear'].value_counts()

# Filter full-length movies (titleType = "movie")
basics = basics[basics['titleType'] == 'movie']

basics['titleType'].value_counts()

# Exclude movies with missing values for genre or runtime
basics = basics.dropna(subset=['genres', 'runtimeMinutes'])

basics['genres'].isna().sum()

basics['runtimeMinutes'].isna().sum()

# Exclude movies from the Documentary genre
is_documentary = basics['genres'].str.contains('documentary', 
                                               case = False)
basics = basics[~is_documentary]

basics.info()

### Title Ratings

keepers = ratings["tconst"].isin(akas["titleId"])
ratings = ratings[keepers]
ratings

ratings.info()

## Save each  file to a compressed csv file

# Save filtered title.basics DataFrame
basics.to_csv('Data/title_basics.csv.gz', compression='gzip', index=False)

# Save filtered title.ratings DataFrame
ratings.to_csv('Data/title_ratings.csv.gz', compression='gzip', index=False)

# Save filtered title.akas DataFrame for US movies
akas.to_csv('Data/title_akas_us.csv.gz', compression='gzip', index=False)


# Open saved file and preview again
basics = pd.read_csv("Data/title_basics.csv.gz", low_memory = False)
basics.info()

# Open saved file and preview again
ratings = pd.read_csv("Data/title_ratings.csv.gz", low_memory = False)
ratings.info()

# Open saved file and preview again
akas = pd.read_csv("Data/title_akas_us.csv.gz", low_memory = False)
akas.info()
