==> Part 3- MySQL Database.ipynb <==
#IMPORTS
import pandas as pd
import numpy as np
from sqlalchemy import create_engine
import pymysql
from urllib.parse import quote_plus as urlquote
import pymysql
pymysql.install_as_MySQLdb()
from sqlalchemy import create_engine
from sqlalchemy_utils import create_database, database_exists
# Create connection string using credentials following this format
#connection = "dialect+driver://username:password@host:port/database"
#connection_str = "mysql+pymysql://root:root@localhost/books"
#connection = f"mysql+pymysql://{login['username']}:{urlquote(login['password'])}@localhost/books"
#engine = create_engine(connection)

import json
with open('/Users/patelmedzy/.secret/mysql.json') as f:
    login = json.load(f)
login.keys()

basics = pd.read_csv('Data/title_basics.csv.gz')
basics.info()

ratings = pd.read_csv('Data/title_ratings.csv.gz')
ratings.info()

tmdb = pd.read_csv('Data/tmdb_results_combined.csv.gz')
tmdb.info()

## Genres

### Getting a List of Unique Genres

## create a col with a list of genres
basics['genres_split'] = basics['genres'].str.split(',')
basics

# Create a new dataframe with a row
# for each genre a movie belongs to by using .explode
exploded_genres = basics.explode('genres_split')
exploded_genres.head(2)

unique_genres = sorted(exploded_genres['genres_split'].unique())
unique_genres


### Create a new title_genres table

title_genres = exploded_genres[['tconst', 'genres_split']].copy()
title_genres.head(2)

### Create a genre mapper dictionary to replace string to genres with integers

# Make a dictionary
# with list of unique genres as the key and the new integer id as values
genre_id_map = dict(zip(unique_genres, range(len(unique_genres))))

genre_id_map

### Replace the string genres in title_genres with the new integer ids

# Create a genre_id column using the genre_map dictionary with .map
title_genres['genre_id'] = title_genres['genres_split'].map(genre_id_map)
# Drop the "genres_split" column
title_genres = title_genres.drop(columns='genres_split')


title_genres.head()

### Convert the genre map dictionary into a dataframe

# Create a dataframe from the genre_map dictionary with .map
genres = pd.DataFrame({'genre_name': genre_id_map.keys(),
                           'genre_id': genre_id_map.values()})

genres.head()

### Saving the MySQL tables with tconst as primary key

# Obtain the MySQL login username and password
with open('/users/patelmedzy/.secret/mysql.json') as f:
    login = json.load(f)
# Display the MySQL login keys
login.keys()


# Create a connection string using credentials following this format:
# connection = "dialect+driver://username:password@host:port/database"
database_name = "Movies"
connection_str = f"mysql+pymysql://{login['username']}:{login['password']}@localhost/{database_name}"

# Create an instance of the sqlalchemy Engine Class 
engine = create_engine(connection_str)


# Create the Movies database, if it does not already exist
if database_exists(connection_str) == False: create_database(connection_str)
else: print('The database already exists.')

### Create SQL Table

#### title_basics SQL Table

basics.head()

basics.info()

# Drop columns from the dataframe
basics = basics.drop(columns = ['titleType', 'originalTitle',
                                                  'isAdult', 'endYear', 'genres', 'genres_split'])

basics.info()

#### Find the max string length

## Example

from sqlalchemy.types import *
## Calculate max string lengths for object columns
key_len = basics['tconst'].fillna('').map(len).max()
title_len = basics['primaryTitle'].fillna('').map(len).max()
## Create a schema dictonary using Sqlalchemy datatype objects
df_schema = {
    "tconst": String(key_len+1), 
    "primaryTitle": Text(title_len+1),
    'startYear':Float(),
    'endYear':Float(),
    'runtimeMinutes':Integer()}



#### Run df.to_sql with the dtype argument.

# Save to sql with dtype and index=False
basics.to_sql('title_basics',
              engine,
              dtype=df_schema,if_exists='replace',
              index=False)



####  Run the query to ADD PRIMARY KEY

engine.execute('ALTER TABLE title_basics ADD PRIMARY KEY (`tconst`);')

# Display the first 5 rows of the table using an SQL query.
q = """
SELECT * 
FROM title_basics LIMIT 5
;"""
pd.read_sql(q, engine)


## title_ratings SQL Table

# Calculate max string lengths for object columns
key_len = ratings['tconst'].fillna('').map(len).max()

# Use 1 + the max_str_len for object columns
# Create a schema dictonary using Sqlalchemy datatype objects
df_schema = {
    "tconst": String(key_len+1), 
    'averageRating':Float(),
    'numVotes':Integer()}

# Save the dataframe to an sql table
# with appropriate datatypes and set index=False
ratings.to_sql('title_ratings',
              engine, 
              dtype=df_schema,
              if_exists='replace',
              index=False)


# Use the sqlalchemy engine to update the table, and 
# set tconst as the primary key
engine.execute('ALTER TABLE title_ratings ADD PRIMARY KEY (`tconst`);')

# Display the first 5 rows of the table using a SQL query
q = """
SELECT * 
FROM title_ratings LIMIT 5
;"""
pd.read_sql(q, engine)


## title_genres SQL Table

# Display the first (5) rows
title_genres.head()


title_genres.info()

# Calculate max string lengths for object columns
tconst_len = title_genres['tconst'].fillna('').map(len).max()

# Use 1 + the max_str_len for object columns
# Create a schema dictonary using Sqlalchemy datatype objects
df_schema = {
    "tconst": String(tconst_len+1), 
    'genre_id':Integer()}


# Save the dataframe to an sql table
# with appropriate datatypes and set index=False
title_genres.to_sql('title_genres',
              engine, 
              dtype=df_schema,
              if_exists='replace',
              index=False)

# Display the first 5 rows of the table using a SQL query
q = """
SELECT * 
FROM title_genres LIMIT 5
;"""
pd.read_sql(q, engine)

# Display the first (5) rows
genres.head()


# Display the names, non-null values, and datatypes for the columns
genres.info()

# Calculate max string lengths for object columns
genre_name_len = genres['genre_name'].fillna('').map(len).max()

# Use 1 + the max_str_len for object columns 
# Create a schema dictonary using Sqlalchemy datatype objects
df_schema = {
    "genre_id": Integer(), 
    'genre_name': String(genre_name_len+1)}

# Save the dataframe to an sql table
# with appropriate datatypes and set index=False
genres.to_sql('genres',
              engine, 
              dtype=df_schema,
              if_exists='replace',
              index=False)

# Use the sqlalchemy engine to update the table and set genre_id as the primary key
engine.execute('ALTER TABLE genres ADD PRIMARY KEY (`genre_id`);')

# Display the first 5 rows of the table using an SQL query
q = """
SELECT * 
FROM genres LIMIT 5
;"""
pd.read_sql(q, engine)

## tmdb SQL Table

tmdb.head()

tmdb.info()

#Drop
tmdb = tmdb.loc[:,["imdb_id","revenue",'budget','certification']]

tmdb.drop(0, inplace=True, axis=0)

tmdb.head(2)

tmdb.info()

# Calculate max string lengths for object columns
imdb_id_len = tmdb['imdb_id'].fillna('').map(len).max()
certification_len = tmdb['certification'].fillna('').map(len).max()


# Use 1 + the max_str_len for object columns
# Create a schema dictonary using Sqlalchemy datatype objects
df_schema = {
    'imdb_id': String(imdb_id_len+1), 
    'budget':Float(),
    'revenue':Float(),
    'certification': String(certification_len+1)}

# Save the dataframe to an sql table
# with appropriate datatypes and set index=False
tmdb.to_sql('tmdb_data',
              engine, 
              dtype=df_schema,
              if_exists='replace',
              index=False)

# Use the sqlalchemy engine to update the table and set imdb_id as the primary key
engine.execute('ALTER TABLE tmdb_data ADD PRIMARY KEY (`imdb_id`);')

# Display the first 5 rows of the table using an SQL query
q = """
SELECT * 
FROM tmdb_data LIMIT 5
;"""
pd.read_sql(q, engine)

## Show SQL Tables

# Display the SQL table names using an SQL query
q = """SHOW TABLES;"""
pd.read_sql(q, engine)
