==> Part 4- Hypothesis Testing.ipynb <==
# Stakeholder Questions
1) Does the MPAA rating of a movie (G/PG/PG-13/R) affect how much revenue the movie generates?
2) Do movies that are over 2.5 hours long earn more revenue than movies that are 1.5 hours long (or less)?
3) Are some genres higher rated than others?


- For each question, they would like:
    - perform a statistical test to get a mathematically-supported answer.
    - report if you found a significant difference between ratings.
    - what was the p-value of your analysis?
    - which rating earns the most revenue?
    - prepare a visualization that supports your finding.



# Import libraries
import numpy as np
import pandas as pd
import glob
import json
import os
import matplotlib.pyplot as plt
from matplotlib.ticker import StrMethodFormatter
price_fmt = StrMethodFormatter("${x:,.0f}")
import seaborn as sns


import tmdbsimple as tmdb
from tqdm.notebook import tqdm_notebook

import pymysql
pymysql.install_as_MySQLdb()
from sqlalchemy.types import *
from sqlalchemy_utils import create_database, database_exists
from sqlalchemy import create_engine

from scipy import stats
## Post Hoc
from statsmodels.stats.multicomp import pairwise_tukeyhsd


#Folder path must end with /
FOLDER = "Data/"
os.makedirs(FOLDER, exist_ok=True)
os.listdir(FOLDER)

# TMDB 

with open('/Users/patelmedzy/.secret/tmdb_api.json', 'r') as f:
    login = json.load(f)
## Display the keys of the loaded dict
login.keys()
tmdb.API_KEY =  login['api-key']

# Load Dataframe
basics = pd.read_csv('Data/title_basics.csv.gz')
basics.info()

def get_movie_with_rating(movie_id):
    #Get the movie object for the current id
    movie = tmdb.Movies(movie_id)
    #Save the .info .releases dictionaries
    info = movie.info()
    releases = movie.releases()
    #Loop thru countries in releases
    for c in releases["countries"]:
        #if the country abbreviation == US
        if c["iso_3166_1"] == "US":
            #Save a certification key to new column
            info["certification"] = c["certification"]
    return info

#Testing API function to get info on "The Avengers" movie
get_movie_with_rating("tt0848228")

## Movies from 2010-2019

YEARS_TO_GET = range(2010,2019)


### Define an errors list

errors = []

### OUTER & INNER Loop

# Start of OUTER loop
for YEAR in tqdm_notebook(YEARS_TO_GET, desc='YEARS', position=0):
    #Defining the JSON file to store results for year
    JSON_FILE = f'{FOLDER}tmdb_api_results_{YEAR}.json'
    # Check if file exists
    file_exists = os.path.isfile(JSON_FILE)
    # If it does not exist: create it
    if file_exists == False:
        # save an empty dict with just "imdb_id" to the new json file.
        with open(JSON_FILE,'w') as f:
            json.dump([{'imdb_id':0}],f)
    #Saving new year as the current df
    df = basics.loc[ basics['startYear']==YEAR].copy()
    # saving movie ids to list
    movie_ids = df['tconst'].copy()
    # Load existing data from json into a dataframe called "previous_df"
    previous_df = pd.read_json(JSON_FILE)
    # filter out any ids that are already in the JSON_FILE
    movie_ids_to_get = movie_ids[~movie_ids.isin(previous_df["imdb_id"])]
    
# Get index and movie id from list
# INNER loop
    for movie_id in tqdm_notebook(movie_ids_to_get,
                                  desc=f'Movies from {YEAR}',
                                  position=1,
                                  leave=True):
        try:
            # Retrieve then data for the movie id
            temp = get_movie_with_rating(movie_id)  
            # Append/extend results to existing file using a pre-made function
            write_json(temp,JSON_FILE)
            # Short 20 ms sleep to prevent overwhelming server
            time.sleep(0.02)
            
        except Exception as e:
            errors.append([movie_id, e])
            
    final_year_df = pd.read_json(JSON_FILE)
    final_year_df.to_csv(f"{FOLDER}final_tmdb_data_{YEAR}.csv.gz", 
                         compression="gzip", index=False)

print(f"- Total errors: {len(errors)}")

### Using glob to get a list of files

import glob
q = "Data/final_tmdb_data_20*.csv.gz"
chunked_files = glob.glob(q)
# Showing the first 5
chunked_files

## Loading all files as df and appending to a list
df_list = []
for file in chunked_files:
    print(file)
    temp_df = pd.read_csv(f"{file}")
    df_list.append(temp_df)


#List comprehension to combine files into 1 df
## Loading and Concatenating the list of dfs with 1 line
df = pd.concat([pd.read_csv(file) for file in chunked_files])
df.info()

# SQL

# Obtain the MySQL login username and password
with open('/Users/patelmedzy/.secret/mysql.json') as f:
    login = json.load(f)
# Display the MySQL login keys
login.keys()

# Create a connection string using credentials following this format:
# connection = "dialect+driver://username:password@host:port/database"
database_name = "Movies"
connection_str = f"mysql+pymysql://{login['username']}:{login['password']}@localhost/{database_name}"

# Create an instance of the sqlalchemy Engine Class using create_engine
engine = create_engine(connection_str)

# Questions 

### Function to interpret p-value

# Create a function to evaluate the p=value of a hypothesis test
def interpret_pvalue(ho, ha, alpha=.05):
    ho_desc = 'No statistical significance exists. The null hypothesis was not rejected.'
    ha_desc = 'A statistical significance exists. The null hypothesis is rejected and the alternative is supported that..'
    if pvalue < alpha:
        print(f'The p-value for the test was {pvalue}')
        print(f'It was < the alpha value of {alpha}, so')
        print(ha_desc)
        print(ha)
    else:
        print(f'The p-value for the test was {pvalue}')
        print(f'It was > the alpha value of {alpha}, so')
        print(ho_desc)
        print(ho)

## 1) Does the MPAA rating of a movie (G/PG/PG-13/R) affect how much revenue the movie generates?


### Stating Hypothesis
- $H_0$ (Null Hypothesis): The MPAA rating of a movie does not affect how much revenue the movie generates.
- $H_A$ (Alternate Hypothesis): The MPAA rating of a movie does affect how much revenue the movie generates.

# Import Data
# Use an SQL query to create a dataframe
q = """
SELECT certification, revenue
FROM tmdb_data
WHERE revenue > 0 AND certification IS NOT NULL
;"""
df = pd.read_sql(q, engine)

df.head()

# Display the unique values and their counts for this column
df['certification'].value_counts()

### Selecting the correct test to perform
- Target Datatype: 
    - Numeric (revenue)
- Number of Samples: 
    - more than 5 samples (certification)
- Test to perform:
    - parametric: ANOVA and/or Tukey
    - non-parametric: Kruskal-Wallis and/or Tukey

### Assumptions
- No significant outliers
- Normality 
- Equal Variance


#### Testing Assumptions: No Significant Outliers

# Creating a loop function to test for number or outliers

groups = {}
for certification in df['certification'].unique():
    cert_df = df.loc[df['certification'] == certification, 'revenue']
    groups[certification] = cert_df
groups.keys()



# Loop through the groups to obtain
# the number of outliers, display the result and then remove them
for certification, revenue in groups.items():
    # Calculate the number of outliers
    outliers = np.abs(stats.zscore(revenue)) > 3
    # Display the number of outliers in the group
    print(f"{outliers.sum()} outliers were removed from the {certification} group.")
    # Remove the outliers
    groups[certification] = revenue.loc[~outliers]

 - No Significant Outliers Assumptions has been met.

#### Testing Assumptions: Normality

# Loop through groups to obtain group count and p-value for Normality test
normality_results = {}
for certification, revenue in groups.items():
    stat, p = stats.normaltest(revenue)
    results[certification] = {'n':len(revenue), 'p':p}

normality_results = pd.DataFrame(normality_results)
normality_results.head()

# transposing the results dataframe
#https://pandas.pydata.org/docs/reference/api/pandas.DataFrame.transpose.html
normality_results = normality_results.T
normality_results.head()

normality_results['sig?'] = normality_results['p'] < .05

normality_results

- The p-values are less than 0.05 for PG-13, PG and R the MPAA ratings. While the p-values for G and NR are greater than 0.05.
- The groups are not large enough to ignore the assumption of Normality. Hence, Assumption for Normality is not met.
- I will perform the non-parametric equivalent of your test: Kruskal-Wallis and/or Tukey

#### Testing Assumption: Equal Variance

# Hypothesis
ho = "All input samples are from populations with equal variances."
ha = "All input samples are not from populations with equal variances."
# Run the test
statistic, pvalue = stats.levene(*groups.values())
# Display the interpretation
interpret_pvalue(ho, ha, alpha=.05)

- interpret results of equal variance test 

### Execute Selected Test (or alternative is test assumptions have not been met)

#### Interpret p-value

# Hypotheses
ho = 'The MPAA rating of a movie does not affect how much revenue the movie generates.'
ha = 'The MPAA rating of a movie does affect how much revenue the movie generates.'
# Run the test
statistic, pvalue = stats.kruskal(*groups.values())
# Display the interpretation
interpret_pvalue(ho, ha, alpha=.05) 

### Show a supporting visualization to display results

fig, ax = plt.subplots(nrows=1, figsize=(8,4), facecolor='w')
plt.title('MPAA rating', fontsize = 22, weight='bold')
sns.barplot(data=df, x='certification', y='revenue', ci=68, palette="viridis")
ax.set_xticklabels(ax.get_xticklabels(), rotation='45', ha='right')
plt.xlabel("MPAA rating", fontsize = 16, weight='bold')
plt.xticks(weight='bold')
ax.set_ylabel('Revenue (dollars)',fontweight='bold',fontsize=14)
ax.set_facecolor('lightblue')
ax.tick_params(labelcolor='k', labelsize=10)
ax.set_yticklabels(ax.get_yticks(), weight='bold')
ax.yaxis.set_major_formatter(price_fmt)
for axis in ['top','bottom','left','right']:
    ax.spines[axis].set_linewidth(3)
plt.tight_layout()
plt.show;
#plt.savefigImages/mages/MPAA Ratings vs Revenue.png');

'''
# Create a box plot
sns.barplot(data=df, x=df['certification'], y=df['revenue'])
plt.xlabel('Ratings')
plt.ylabel('Revenue')
plt.title('Ratings Revenue')
plt.xticks(rotation=45)
plt.show()
'''

## 2) Do movies that are over 2.5 hours long earn more revenue than movies that are 1.5 hours long (or less)?

### Stating Hypothesis
- $H_0$ (Null Hypothesis): No difference. Movies that are over 2.5 hours do not have a significantly different revenue than movies that under 1.5 hours in length.
- $H_A$ (Alternate Hypothesis): There is a significant differnce. Movies that are over 2.5 hours have a significantly different revenue than movies that under 1.5 hours in length.

### Selecting the correct test to perform
- Target Datatype: 
    - Numeric (revenue)
- Number of Samples: 
    - 2 samples (movies less than 1.5 hours in length, and movies greater than 2.5 hours in length)
- Test to perform:
    - parametric: 2 Sample T-Test
    - non-parametric: Welch's T-Test

### Assumptions
- No significant outliers
- Normality 
- Equal Variance


## 3) Are some genres higher rated than others?

### Stating Hypothesis
- $H_0$ (Null Hypothesis): No difference. The genre of a movei does not affect the amount of revenue generated by a movie.
- $H_A$ (Alternate Hypothesis): There is a significant difference. The genre of a movie does affect the amount of revenue generated by a movie.

### Selecting the correct test to perform
- Target Datatype: 
    - Numeric (revenue)
- Number of Samples: 
    - more than 2 samples (genres)
- Test to perform:
    - parametric: ANOVA and/or Tukey
    - non-parametric: Kruskal-Wallis and/or Tukey

### Assumptions
- No significant outliers
- Normality 
- Equal Variance
